+++ db/migrations/schemas/ingredient-cost.js
@@ -1,5 +1,5 @@
 module.exports = (table) => {
   table.increments()
-  table.integer('cost')
+  table.decimal('cost')
   table.string('wholesaler')
 }
